Tests work at all.
The mocking of prisma isn't working correctly yet. So you have to have a running mysql instance to run the tests.
Also there are some things around the lastActive timestamp stuff that I don't like.
Because it's in the schema jest or ts is making us include it when createing a user even though it has a default value.
This makes the expects more complicated than they need to be :-(